refactor: relocate specs to column and add re-export
